Crregido erro de decimal

+++ back/database/migrations/2026_01_17_112555_create_orders_table.php
@@ -19,7 +19,7 @@ return new class extends Migration {
             $table->string('localizador')->nullable();
 
             // Totales (en centavos)
-            $table->integer('amount_total')->default(0);
+            $table->decimal('amount_total', 10, 2)->default(0);
             $table->string('currency', 10)->default('eur');
 
             // Estado
